+++ RELEASE_NOTES.rst
@@ -13,6 +13,7 @@ FACET 2.0
 - API: class :class:`.RangePartitioner` supports new optional arguments ``lower_bound``
   and ``upper_bound`` in method :meth:`~.RangePartitioner.fit` and no longer accepts
   them in the class initializer
+- API: rename method ``Sample.subsample()`` to :meth:`.Sample.resample`
 
 ``facet.inspection``
 ^^^^^^^^^^^^^^^^^^^^